ProcessB_ThreadManagement.c
Contains the manageThreads function, which takes a void* parameter and returns void* (which will be a pointer to the thread). Currently testing with a dummy function called fnName, which the thread is supposed to execute. Working on making the thread persist as long as Process B runs. Having issues using a dynamically allocated thread with pthread_create().